use handle hook instead of user defined endpoints
